requirements.txt
kafka-python==2.3.0
websocket-client==1.9.0
pyspark==3.5.7
